services/backend/requirements.txt
aerich==0.7.1
aiokafka==0.8.0
aiosqlite==0.17.0
aiofiles==23.1.0
anyio==3.6.2
asyncpg==0.27.0
attrs==22.2.0
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
dictdiffer==0.9.0
debugpy==1.6.7
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.4.8
pydantic==1.10.7
pypika-tortoise==0.1.6
pytest==7.2.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2023.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
tomlkit==0.11.7
tortoise-orm==0.19.2
typing_extensions==4.5.0
uvicorn[standard]==0.20.0
aiopg==1.4.0